+++ src/app/users/components/dynamic-input.component.html
@@ -10,7 +10,7 @@
   />
 
   <div class="actions">
-    <button mat-icon-button (click)="onSubmit(value())">
+    <button mat-icon-button (click)="onSubmit()">
       <mat-icon>check</mat-icon>
     </button>
     <button mat-icon-button (click)="onClose()">
